backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib==1.7.4
openai==2.31.0
pypdf==6.0.0
python-docx==1.2.0
python-pptx==1.0.2
email-validator==2.3.0
